+++ hlp/chat/.gitignore
@@ -9,6 +9,7 @@ __pycache__/
 # Distribution / packaging
 .Python
 data/
+checkpoint/
 build/
 develop-eggs/
 dist/
