+++ hlp/chat/README.md
@@ -1,8 +1,11 @@
 # 运行说明
-执行： python execute.py [模型] [运行参数]
++ execute.py主入口文件：指令需要附带运行参数
++ cmd：python execute.py [模型类别] [执行模式]
++ 模型类别：seq2seq/gpt2
++ 执行类别：chat/train
 
-例如： python execute.py seq2seq train，即为训练seq2seq模型
-再如： python execute.py seq2seq chat，即为运行对话系统
++ chat模式为对话模式。chat模式下运行时，输入exit即退出对话
++ train模式为训练模式，一般建议在首次运行时，先执行train模式
 
 **相关参数后继补充**
