+++ hlp/chat/config/getConfig.py
@@ -1,22 +1,30 @@
-from configparser import ConfigParser
+import os
 import json
-import xml.dom.minidom as xl
+
+seq2seq_config = os.path.dirname(__file__) + r'\seq2seq.json'
+path = os.path.dirname(__file__)[:-6]
 
 
 def get_config_json(config_file='main.json'):
-    with open(config_file) as file:
+    with open(config_file, 'r') as file:
         return json.load(file)
 
 
-def get_config_xml(config_file='main.xml'):
-    dom = xl.parse(config_file)
-    return dom.documentElement
+def config(config_file=seq2seq_config):
+    return get_config_json(config_file=config_file)
 
 
-def get_config_ini(config_file='main.ini'):
-    parser = ConfigParser()
-    parser.read(config_file)
-    _conf_ints = [(key, int(value)) for key, value in parser.items('ints')]
-    _conf_floats = [(key, float(value)) for key, value in parser.items('floats')]
-    _conf_strings = [(key, str(value)) for key, value in parser.items('strings')]
-    return dict(_conf_ints + _conf_floats + _conf_strings)
+conf = {}
+
+conf = config()
+
+vocab_inp_size = conf['enc_vocab_size']
+vocab_tar_size = conf['dec_vocab_size']
+embedding_dim = conf['embedding_dim']
+BATCH_SIZE = conf['batch_size']
+units = conf['layer_size']
+max_length_inp = 20
+max_length_tar = 20
+max_train_data_size = conf['max_train_data_size']
+data = path + conf['seq_data']
+train_data = path + conf['train_data']
