+++ hlp/chat/execute.py
@@ -1,15 +1,32 @@
 import sys
-import config.getConfig as getConfig
-import model.Seq2Seq.seq2seq as seq2seq
+import os
+from model.Seq2Seq.trainer import train
+from model.Seq2Seq.predict import predict
+
+'''
+主入口：指令需要附带运行参数
+cmd：python execute.py [模型类别] [执行模式]
+模型类别：seq2seq/gpt2
+执行类别：chat/train
+
+chat模式下运行时，输入exit即退出对话
+'''
 
 if __name__ == '__main__':
-    if sys.argv[1] == 'seq2seq':
+    if len(sys.argv) != 3:
+        print('Error:指令执行有误，详细见执行说明!')
+
+    elif sys.argv[1] == 'seq2seq':
         if sys.argv[2] == 'train':
-            seq2seq.train()
+            train()
         elif sys.argv[2] == 'chat':
             while (True):
                 sentence = input('User:')
                 if sentence == 'exit':
                     break
                 else:
-                    print('ChatBot:', seq2seq.predict(sentence))
+                    print('ChatBot:', predict(sentence))
+        else:
+            print('Error:不存在', sys.argv[2], '模式!')
+    else:
+        print('Error:不存在', sys.argv[1], '模型类别!')
